+++ CHANGES.rst
@@ -5,8 +5,12 @@ ddbmock 0.3.2.dev
 This section documents all user visible changes included between ddbmock
 versions 0.3.1 and versions 0.3.2
 
-Additions
----------
+Changes
+-------
+
+- fix number validation
+- return no Item field when not found
+
 
 =============
 ddbmock 0.3.1